+++ packages/asc-ui/src/utils/index.ts
@@ -1,4 +1,3 @@
-import { getValueFromTheme } from './themeUtils'
 import deprecatedWarning from './deprecatedWarning'
 
 export {
@@ -10,9 +9,6 @@ export {
   showHide,
   ShowHideTypes,
   showAboveBackDrop,
-  customCss,
-  CustomCssPropsType,
-  CustomCssPropsInterface,
   getValueFromTheme,
   themeSpacing,
   calculateFluidStyle,
@@ -20,9 +16,4 @@ export {
 export { valueFromObject } from './valueFromObject'
 export { fromProps } from './fromProps'
 
-/**
- * @deprecated Use getValueFromTheme instead
- */
-const fromTheme = getValueFromTheme
-
-export { fromTheme, deprecatedWarning }
+export { deprecatedWarning }